requirements.txt
PyQt6
requests
python-dotenv
gkeepapi
google-api-python-client